fix(bme680): correct multi-language bugs found during hardware verification
Python: float//int gives float which cannot be bit-shifted in
_calc_heater_resistance; fix by wrapping with int().

C++: uint8_t _heat_temp truncates 320 to 64; fix by using int16_t.

Rust: compensate_humidity overflows i32 with temp_scaled*par_h5/100
intermediate; fix by promoting all humidity locals to i64.

Node.js: >> applies ToInt32 before shift, corrupting intermediates
>2^31 in _compensateHumidity and _compensateGas; fix by replacing
affected shifts with Math.floor(x/2^n).

Kotlin: signExtend8() was called on values already sign-extended by
Byte.toInt(), producing results 256 too negative for any negative
calibration byte; removing the redundant call fixes parT3, parP3,
parP6, parP7, parH3-H5, parH7, parG1, parG3, resHeatVal.

JVM examples: add I2C_BUS/I2C_ADDR env-var support; fix infinite
while(true) loops in Minimal examples; fix wrong require() paths in
Node.js examples.

All bugs verified against hardware (BME680 on i2c-1 at 0x77).

+++ cpp/src/chips/environmental/BME680.cpp
@@ -255,7 +255,7 @@ void BME680Full::set_filter(uint8_t coeff) {
 }
 
 void BME680Full::set_heater(int16_t temp_c, uint16_t duration_ms) {
-    _heat_temp = (uint8_t)temp_c;
+    _heat_temp = temp_c;
     _heat_dur = duration_ms;
     _setup_heater(0, temp_c, duration_ms);
     _write_reg(REG_CTRL_GAS_1, (1 << 4) | 0);
